reworking Integer and Copy
