+++ hacspec/src/machine_integers.rs
@@ -109,7 +109,7 @@ macro_rules! implement_public_mi {
     ($t:ty,$bits:literal,$true_val:expr) => {
         impl NumericCopy for $t {}
         impl Integer for $t {
-            const NUM_BITS: u32 = $bits;
+            const NUM_BITS: usize = $bits;
             
             #[inline]
             #[cfg_attr(feature = "use_attributes", library(hacspec))]
@@ -373,7 +373,7 @@ macro_rules! implement_secret_mi {
     ($t:ident,$base:ty,$bits:literal) => {
         impl NumericCopy for $t {}
         impl Integer for $t {
-            const NUM_BITS: u32 = $bits;
+            const NUM_BITS: usize = $bits;
 
             #[inline]
             #[cfg_attr(feature = "use_attributes", library(hacspec))]
@@ -527,3 +527,329 @@ implement_secret_signed_mi!(I16, i16, 16);
 implement_secret_signed_mi!(I32, i32, 32);
 implement_secret_signed_mi!(I64, i64, 64);
 implement_secret_signed_mi!(I128, i128, 128);
+
+impl UnsignedPublicInteger for u8 {
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_le_bytes(self) -> Seq<u8>{
+        let mut x = Seq::new(1);
+        x[0] = self;
+        x
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_be_bytes(self) -> Seq<u8> {
+        let mut x = Seq::new(1);
+        x[0] = self;
+        x
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_le_bytes(x: &Seq<u8>) -> Self {
+        assert!(x.len() == 1);
+        x[0]
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_be_bytes(x: &Seq<u8>) -> Self {
+        assert!(x.len() == 1);
+        x[0]
+    }
+}
+
+impl UnsignedPublicInteger for u16 {
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_le_bytes(self) -> Seq<u8> {
+        Seq::from_seq(&u16_to_le_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_be_bytes(self) -> Seq<u8> {
+        Seq::from_seq(&u16_to_be_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_le_bytes(x: &Seq<u8>) -> Self {
+        u16_from_le_bytes(u16Word::from_seq(x))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_be_bytes(x: &Seq<u8>) -> Self {
+        u16_from_be_bytes(u16Word::from_seq(x))
+    }
+}
+
+impl UnsignedPublicInteger for u32 {
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_le_bytes(self) -> Seq<u8> {
+        Seq::from_seq(&u32_to_le_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_be_bytes(self) -> Seq<u8> {
+        Seq::from_seq(&u32_to_be_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_le_bytes(x: &Seq<u8>) -> Self {
+        u32_from_le_bytes(u32Word::from_seq(x))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_be_bytes(x: &Seq<u8>) -> Self {
+        u32_from_be_bytes(u32Word::from_seq(x))
+    }
+}
+
+impl UnsignedPublicInteger for u64 {
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_le_bytes(self) -> Seq<u8> {
+        Seq::from_seq(&u64_to_le_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_be_bytes(self) -> Seq<u8> {
+        Seq::from_seq(&u64_to_be_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_le_bytes(x: &Seq<u8>) -> Self {
+        u64_from_le_bytes(u64Word::from_seq(x))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_be_bytes(x: &Seq<u8>) -> Self {
+        u64_from_be_bytes(u64Word::from_seq(x))
+    }
+}
+
+impl UnsignedPublicInteger for u128 {
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_le_bytes(self) -> Seq<u8> {
+        Seq::from_seq(&u128_to_le_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_be_bytes(self) -> Seq<u8> {
+        Seq::from_seq(&u128_to_be_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_le_bytes(x: &Seq<u8>) -> Self {
+        u128_from_le_bytes(u128Word::from_seq(x))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_be_bytes(x: &Seq<u8>) -> Self {
+        u128_from_be_bytes(u128Word::from_seq(x))
+    }
+}
+
+impl UnsignedInteger for U8 {}
+impl UnsignedInteger for U16 {}
+impl UnsignedInteger for U32 {}
+impl UnsignedInteger for U64 {}
+impl UnsignedInteger for U128 {}
+impl UnsignedInteger for u8 {}
+impl UnsignedInteger for u16 {}
+impl UnsignedInteger for u32 {}
+impl UnsignedInteger for u64 {}
+impl UnsignedInteger for u128 {}
+
+impl SignedInteger for I8 {}
+impl SignedInteger for I16 {}
+impl SignedInteger for I32 {}
+impl SignedInteger for I64 {}
+impl SignedInteger for I128 {}
+impl SignedInteger for i8 {}
+impl SignedInteger for i16 {}
+impl SignedInteger for i32 {}
+impl SignedInteger for i64 {}
+impl SignedInteger for i128 {}
+
+impl PublicInteger for u8 {
+    type SecretVersion = U8;
+}
+impl PublicInteger for u16 {
+    type SecretVersion = U16;
+}
+impl PublicInteger for u32 {
+    type SecretVersion = U32;
+}
+impl PublicInteger for u64 {
+    type SecretVersion = U64;
+}
+impl PublicInteger for u128 {
+    type SecretVersion = U128;
+}
+impl PublicInteger for i8 {
+    type SecretVersion = I8;
+}
+impl PublicInteger for i16 {
+    type SecretVersion = I16;
+}
+impl PublicInteger for i32 {
+    type SecretVersion = I32;
+}
+impl PublicInteger for i64 {
+    type SecretVersion = I64;
+}
+impl PublicInteger for i128 {
+    type SecretVersion = I128;
+}
+
+impl SecretInteger for U8 {
+    type PublicVersion = u8;
+    #[cfg_attr(feature = "use_attributes", library(hacspec))]
+    fn classify(x: Self::PublicVersion) -> Self {
+        U8(x)
+    }
+}
+impl SecretInteger for U16 {
+    type PublicVersion = u16;
+    #[cfg_attr(feature = "use_attributes", library(hacspec))]
+    fn classify(x: Self::PublicVersion) -> Self {
+        U16(x)
+    }
+}
+impl SecretInteger for U32 {
+    type PublicVersion = u32;
+    #[cfg_attr(feature = "use_attributes", library(hacspec))]
+    fn classify(x: Self::PublicVersion) -> Self {
+        U32(x)
+    }
+}
+impl SecretInteger for U64 {
+    type PublicVersion = u64;
+    #[cfg_attr(feature = "use_attributes", library(hacspec))]
+    fn classify(x: Self::PublicVersion) -> Self {
+        U64(x)
+    }
+}
+impl SecretInteger for U128 {
+    type PublicVersion = u128;
+    #[cfg_attr(feature = "use_attributes", library(hacspec))]
+    fn classify(x: Self::PublicVersion) -> Self {
+        U128(x)
+    }
+}
+impl SecretInteger for I8 {
+    type PublicVersion = i8;
+    #[cfg_attr(feature = "use_attributes", library(hacspec))]
+    fn classify(x: Self::PublicVersion) -> Self {
+        I8(x)
+    }
+}
+impl SecretInteger for I16 {
+    type PublicVersion = i16;
+    #[cfg_attr(feature = "use_attributes", library(hacspec))]
+    fn classify(x: Self::PublicVersion) -> Self {
+        I16(x)
+    }
+}
+impl SecretInteger for I32 {
+    type PublicVersion = i32;
+    #[cfg_attr(feature = "use_attributes", library(hacspec))]
+    fn classify(x: Self::PublicVersion) -> Self {
+        I32(x)
+    }
+}
+impl SecretInteger for I64 {
+    type PublicVersion = i64;
+    #[cfg_attr(feature = "use_attributes", library(hacspec))]
+    fn classify(x: Self::PublicVersion) -> Self {
+        I64(x)
+    }
+}
+impl SecretInteger for I128 {
+    type PublicVersion = i128;
+    #[cfg_attr(feature = "use_attributes", library(hacspec))]
+    fn classify(x: Self::PublicVersion) -> Self {
+        I128(x)
+    }
+}
+
+impl UnsignedSecretInteger for U8 {
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_le_bytes(self) -> Seq<U8>{
+        let mut x = Seq::new(1);
+        x[0] = self;
+        x
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_be_bytes(self) -> Seq<U8> {
+        let mut x = Seq::new(1);
+        x[0] = self;
+        x
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_le_bytes(x: &Seq<U8>) -> Self {
+        assert!(x.len() == 1);
+        x[0]
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_be_bytes(x: &Seq<U8>) -> Self {
+        assert!(x.len() == 1);
+        x[0]
+    }
+}
+
+impl UnsignedSecretInteger for U16 {
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_le_bytes(self) -> Seq<U8> {
+        Seq::from_seq(&U16_to_le_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_be_bytes(self) -> Seq<U8> {
+        Seq::from_seq(&U16_to_be_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_le_bytes(x: &Seq<U8>) -> Self {
+        U16_from_le_bytes(U16Word::from_seq(x))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_be_bytes(x: &Seq<U8>) -> Self {
+        U16_from_be_bytes(U16Word::from_seq(x))
+    }
+}
+
+impl UnsignedSecretInteger for U32 {
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_le_bytes(self) -> Seq<U8> {
+        Seq::from_seq(&U32_to_le_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_be_bytes(self) -> Seq<U8> {
+        Seq::from_seq(&U32_to_be_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_le_bytes(x: &Seq<U8>) -> Self {
+        U32_from_le_bytes(U32Word::from_seq(x))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_be_bytes(x: &Seq<U8>) -> Self {
+        U32_from_be_bytes(U32Word::from_seq(x))
+    }
+}
+
+impl UnsignedSecretInteger for U64 {
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_le_bytes(self) -> Seq<U8> {
+        Seq::from_seq(&U64_to_le_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_be_bytes(self) -> Seq<U8> {
+        Seq::from_seq(&U64_to_be_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_le_bytes(x: &Seq<U8>) -> Self {
+        U64_from_le_bytes(U64Word::from_seq(x))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_be_bytes(x: &Seq<U8>) -> Self {
+        U64_from_be_bytes(U64Word::from_seq(x))
+    }
+}
+
+impl UnsignedSecretInteger for U128 {
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_le_bytes(self) -> Seq<U8> {
+        Seq::from_seq(&U128_to_le_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn to_be_bytes(self) -> Seq<U8> {
+        Seq::from_seq(&U128_to_be_bytes(self))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_le_bytes(x: &Seq<U8>) -> Self {
+        U128_from_le_bytes(U128Word::from_seq(x))
+    }
+    #[cfg_attr(feature="use_attributes", library(hacspec))]
+    fn from_be_bytes(x: &Seq<U8>) -> Self {
+        U128_from_be_bytes(U128Word::from_seq(x))
+    }
+}
