+++ hacspec/src/prelude.rs
@@ -5,6 +5,7 @@
 //! ```
 
 pub use crate::array::*;
+pub use crate::bigint_integers::*;
 pub use crate::seq::*;
 pub use crate::util::*;
 pub use crate::machine_integers::*;
@@ -23,8 +24,9 @@ pub use secret_integers::*;
 #[cfg(feature = "use_attributes")]
 pub use hacspec_attributes::*;
 
-pub use num::{self, BigUint, CheckedSub, Num, Zero};
+pub use num::{self, BigUint, CheckedSub, Num, Zero, traits::sign::Signed};
 pub use rand::{self, Rng, thread_rng};
 pub use std::num::ParseIntError;
 pub use std::ops::*;
 pub use std::{cmp::min, cmp::PartialEq, fmt, fmt::Debug};
+pub use std::str::FromStr;
