+++ hacspec/src/traits.rs
@@ -69,7 +69,7 @@ pub trait SeqTrait<T: Copy>:
 /// types. It offers bit manipulation, instantiation from literal, and convenient
 /// constants.
 pub trait Integer: Numeric {
-    const NUM_BITS: u32;
+    const NUM_BITS: usize;
 
     // Some useful values.
     // Not constants because math integers can't do that.
@@ -86,7 +86,7 @@ pub trait Integer: Numeric {
     /// Get bit `i` of this integer.
     #[inline]
     #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn get_bit(self, i: u32) -> Self {
+    fn get_bit(self, i: usize) -> Self {
         (self >> i) & Self::ONE()
     }
 
@@ -94,7 +94,7 @@ pub trait Integer: Numeric {
     /// Bit `b` has to be `0` or `1`.
     #[inline]
     #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn set_bit(self, b: Self, i: u32) -> Self {
+    fn set_bit(self, b: Self, i: usize) -> Self {
         debug_assert!(b.clone().equal(Self::ONE()) || b.clone().equal(Self::ZERO()));
         let tmp1 = Self::from_literal(!(1 << i));
         let tmp2 = b << i;
@@ -104,162 +104,48 @@ pub trait Integer: Numeric {
     /// Set bit `pos` of this integer to bit `yi` of integer `y`.
     #[inline]
     #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn set(self, pos: u32, y: Self, yi: u32) -> Self {
+    fn set(self, pos: usize, y: Self, yi: usize) -> Self {
         let b = y.get_bit(yi);
         self.set_bit(b, pos)
     }
 
     #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn rotate_left(self, n: u32) -> Self {
+    fn rotate_left(self, n: usize) -> Self {
         // Taken from https://blog.regehr.org/archives/1063
         assert!(n < Self::NUM_BITS);
-        (self.clone() << n) | (self >> ((-(n as i32) as u32) & (Self::NUM_BITS - 1)))
+        (self.clone() << n) | (self >> ((-(n as i32) as usize) & (Self::NUM_BITS - 1)))
     }
 
     #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn rotate_right(self, n: u32) -> Self {
+    fn rotate_right(self, n: usize) -> Self {
         // Taken from https://blog.regehr.org/archives/1063
         assert!(n < Self::NUM_BITS);
-        (self.clone() >> n) | (self << ((-(n as i32) as u32) & (Self::NUM_BITS - 1)))
+        (self.clone() >> n) | (self << ((-(n as i32) as usize) & (Self::NUM_BITS - 1)))
     }
 }
 
 pub trait SecretInteger: Integer {
-    type PublicVersion : PublicInteger;
+    type PublicVersion : PublicInteger + Copy;
     fn classify(x: Self::PublicVersion) -> Self;
 }
-
-impl SecretInteger for U8 {
-    type PublicVersion = u8;
-    #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn classify(x: Self::PublicVersion) -> Self {
-        U8(x)
-    }
-}
-impl SecretInteger for U16 {
-    type PublicVersion = u16;
-    #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn classify(x: Self::PublicVersion) -> Self {
-        U16(x)
-    }
-}
-impl SecretInteger for U32 {
-    type PublicVersion = u32;
-    #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn classify(x: Self::PublicVersion) -> Self {
-        U32(x)
-    }
-}
-impl SecretInteger for U64 {
-    type PublicVersion = u64;
-    #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn classify(x: Self::PublicVersion) -> Self {
-        U64(x)
-    }
-}
-impl SecretInteger for U128 {
-    type PublicVersion = u128;
-    #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn classify(x: Self::PublicVersion) -> Self {
-        U128(x)
-    }
-}
-impl SecretInteger for I8 {
-    type PublicVersion = i8;
-    #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn classify(x: Self::PublicVersion) -> Self {
-        I8(x)
-    }
-}
-impl SecretInteger for I16 {
-    type PublicVersion = i16;
-    #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn classify(x: Self::PublicVersion) -> Self {
-        I16(x)
-    }
-}
-impl SecretInteger for I32 {
-    type PublicVersion = i32;
-    #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn classify(x: Self::PublicVersion) -> Self {
-        I32(x)
-    }
-}
-impl SecretInteger for I64 {
-    type PublicVersion = i64;
-    #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn classify(x: Self::PublicVersion) -> Self {
-        I64(x)
-    }
-}
-impl SecretInteger for I128 {
-    type PublicVersion = i128;
-    #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn classify(x: Self::PublicVersion) -> Self {
-        I128(x)
-    }
-}
+// impl SecretInteger for BigInt {
+//     type PublicVersion = BigInt;
+//     #[cfg_attr(feature = "use_attributes", library(hacspec))]
+//     fn classify(x: Self::PublicVersion) -> Self {
+//         x
+//     }
+// }
 
 pub trait PublicInteger: Integer {
-    type SecretVersion : Integer;
+    type SecretVersion : Integer + Copy;
 }
-
-impl PublicInteger for u8 {
-    type SecretVersion = U8;
-}
-impl PublicInteger for u16 {
-    type SecretVersion = U16;
-}
-impl PublicInteger for u32 {
-    type SecretVersion = U32;
-}
-impl PublicInteger for u64 {
-    type SecretVersion = U64;
-}
-impl PublicInteger for u128 {
-    type SecretVersion = U128;
-}
-impl PublicInteger for i8 {
-    type SecretVersion = I8;
-}
-impl PublicInteger for i16 {
-    type SecretVersion = I16;
-}
-impl PublicInteger for i32 {
-    type SecretVersion = I32;
-}
-impl PublicInteger for i64 {
-    type SecretVersion = I64;
-}
-impl PublicInteger for i128 {
-    type SecretVersion = I128;
-}
+// impl PublicInteger for BigInt {
+//     type SecretVersion = BigInt;
+// }
 
 pub trait UnsignedInteger: Integer {}
 
-impl UnsignedInteger for U8 {}
-impl UnsignedInteger for U16 {}
-impl UnsignedInteger for U32 {}
-impl UnsignedInteger for U64 {}
-impl UnsignedInteger for U128 {}
-impl UnsignedInteger for u8 {}
-impl UnsignedInteger for u16 {}
-impl UnsignedInteger for u32 {}
-impl UnsignedInteger for u64 {}
-impl UnsignedInteger for u128 {}
-
 pub trait SignedInteger: Integer {}
-
-impl SignedInteger for I8 {}
-impl SignedInteger for I16 {}
-impl SignedInteger for I32 {}
-impl SignedInteger for I64 {}
-impl SignedInteger for I128 {}
-impl SignedInteger for i8 {}
-impl SignedInteger for i16 {}
-impl SignedInteger for i32 {}
-impl SignedInteger for i64 {}
-impl SignedInteger for i128 {}
 
 pub trait UnsignedSecretInteger : UnsignedInteger + SecretInteger {
     fn to_le_bytes(self) -> Seq<U8>;
@@ -269,109 +155,8 @@ pub trait UnsignedSecretInteger : UnsignedInteger + SecretInteger {
     /// Get byte `i` of this integer.
     #[inline]
     #[cfg_attr(feature = "use_attributes", library(hacspec))]
-    fn get_byte(self, i: u32) -> Self {
+    fn get_byte(self, i: usize) -> Self {
         (self >> (i * 8)) & ((Self::ONE() << 8) - Self::ONE())
-    }
-}
-
-impl UnsignedSecretInteger for U8 {
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_le_bytes(self) -> Seq<U8>{
-        let mut x = Seq::new(1);
-        x[0] = self;
-        x
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_be_bytes(self) -> Seq<U8> {
-        let mut x = Seq::new(1);
-        x[0] = self;
-        x
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_le_bytes(x: &Seq<U8>) -> Self {
-        assert!(x.len() == 1);
-        x[0]
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_be_bytes(x: &Seq<U8>) -> Self {
-        assert!(x.len() == 1);
-        x[0]
-    }
-}
-
-impl UnsignedSecretInteger for U16 {
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_le_bytes(self) -> Seq<U8> {
-        Seq::from_seq(&U16_to_le_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_be_bytes(self) -> Seq<U8> {
-        Seq::from_seq(&U16_to_be_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_le_bytes(x: &Seq<U8>) -> Self {
-        U16_from_le_bytes(U16Word::from_seq(x))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_be_bytes(x: &Seq<U8>) -> Self {
-        U16_from_be_bytes(U16Word::from_seq(x))
-    }
-}
-
-impl UnsignedSecretInteger for U32 {
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_le_bytes(self) -> Seq<U8> {
-        Seq::from_seq(&U32_to_le_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_be_bytes(self) -> Seq<U8> {
-        Seq::from_seq(&U32_to_be_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_le_bytes(x: &Seq<U8>) -> Self {
-        U32_from_le_bytes(U32Word::from_seq(x))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_be_bytes(x: &Seq<U8>) -> Self {
-        U32_from_be_bytes(U32Word::from_seq(x))
-    }
-}
-
-impl UnsignedSecretInteger for U64 {
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_le_bytes(self) -> Seq<U8> {
-        Seq::from_seq(&U64_to_le_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_be_bytes(self) -> Seq<U8> {
-        Seq::from_seq(&U64_to_be_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_le_bytes(x: &Seq<U8>) -> Self {
-        U64_from_le_bytes(U64Word::from_seq(x))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_be_bytes(x: &Seq<U8>) -> Self {
-        U64_from_be_bytes(U64Word::from_seq(x))
-    }
-}
-
-impl UnsignedSecretInteger for U128 {
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_le_bytes(self) -> Seq<U8> {
-        Seq::from_seq(&U128_to_le_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_be_bytes(self) -> Seq<U8> {
-        Seq::from_seq(&U128_to_be_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_le_bytes(x: &Seq<U8>) -> Self {
-        U128_from_le_bytes(U128Word::from_seq(x))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_be_bytes(x: &Seq<U8>) -> Self {
-        U128_from_be_bytes(U128Word::from_seq(x))
     }
 }
 
@@ -380,107 +165,6 @@ pub trait UnsignedPublicInteger : UnsignedInteger + PublicInteger {
     fn to_be_bytes(self) -> Seq<u8>;
     fn from_le_bytes(x: &Seq<u8>) -> Self;
     fn from_be_bytes(x: &Seq<u8>) -> Self;
-}
-
-impl UnsignedPublicInteger for u8 {
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_le_bytes(self) -> Seq<u8>{
-        let mut x = Seq::new(1);
-        x[0] = self;
-        x
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_be_bytes(self) -> Seq<u8> {
-        let mut x = Seq::new(1);
-        x[0] = self;
-        x
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_le_bytes(x: &Seq<u8>) -> Self {
-        assert!(x.len() == 1);
-        x[0]
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_be_bytes(x: &Seq<u8>) -> Self {
-        assert!(x.len() == 1);
-        x[0]
-    }
-}
-
-impl UnsignedPublicInteger for u16 {
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_le_bytes(self) -> Seq<u8> {
-        Seq::from_seq(&u16_to_le_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_be_bytes(self) -> Seq<u8> {
-        Seq::from_seq(&u16_to_be_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_le_bytes(x: &Seq<u8>) -> Self {
-        u16_from_le_bytes(u16Word::from_seq(x))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_be_bytes(x: &Seq<u8>) -> Self {
-        u16_from_be_bytes(u16Word::from_seq(x))
-    }
-}
-
-impl UnsignedPublicInteger for u32 {
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_le_bytes(self) -> Seq<u8> {
-        Seq::from_seq(&u32_to_le_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_be_bytes(self) -> Seq<u8> {
-        Seq::from_seq(&u32_to_be_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_le_bytes(x: &Seq<u8>) -> Self {
-        u32_from_le_bytes(u32Word::from_seq(x))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_be_bytes(x: &Seq<u8>) -> Self {
-        u32_from_be_bytes(u32Word::from_seq(x))
-    }
-}
-
-impl UnsignedPublicInteger for u64 {
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_le_bytes(self) -> Seq<u8> {
-        Seq::from_seq(&u64_to_le_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_be_bytes(self) -> Seq<u8> {
-        Seq::from_seq(&u64_to_be_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_le_bytes(x: &Seq<u8>) -> Self {
-        u64_from_le_bytes(u64Word::from_seq(x))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_be_bytes(x: &Seq<u8>) -> Self {
-        u64_from_be_bytes(u64Word::from_seq(x))
-    }
-}
-
-impl UnsignedPublicInteger for u128 {
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_le_bytes(self) -> Seq<u8> {
-        Seq::from_seq(&u128_to_le_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn to_be_bytes(self) -> Seq<u8> {
-        Seq::from_seq(&u128_to_be_bytes(self))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_le_bytes(x: &Seq<u8>) -> Self {
-        u128_from_le_bytes(u128Word::from_seq(x))
-    }
-    #[cfg_attr(feature="use_attributes", library(hacspec))]
-    fn from_be_bytes(x: &Seq<u8>) -> Self {
-        u128_from_be_bytes(u128Word::from_seq(x))
-    }
 }
 
 pub trait ModNumeric {
@@ -511,8 +195,8 @@ pub trait Numeric:
     + BitXor<Self, Output = Self>
     + BitOr<Self, Output = Self>
     + BitAnd<Self, Output = Self>
-    + Shl<u32, Output = Self>
-    + Shr<u32, Output = Self>
+    + Shl<usize, Output = Self>
+    + Shr<usize, Output = Self>
     + Not<Output = Self>
     + Default
     + Clone
